backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
google-genai==0.8.0
redis==5.0.1
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic-settings==2.0.3
